[W06H01] Moved classes up one folder to prevent confusion for people already using the previous tests
