+++ w06h01/test/pgdp/ZipTest.java
@@ -1,4 +1,4 @@
-package pgdp.datastructures;
+package pgdp;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
